Package the toolkit and expose a `morphometrics` CLI
Convert the loose root scripts into an installable `surface_morphometrics`
package driven by a single grouped `morphometrics` command.

- Move core modules into surface_morphometrics/ and make intra-package
  imports relative; add pyproject.toml with a `morphometrics` entry point
  and config_template.yml as package data.
- cli.py: lazy click group (subcommands import their module only when
  invoked, so --help and shell completion work without graph-tool/pymeshlab/
  pycurv), grouped/ordered help, per-command "next step" hints, and a
  `new_config` command that writes a starter config.yml.
- Convert the three argv/top-level scripts (make_meshes, pycurv,
  distances_orientations) to click commands; route the xyz2ply subprocess
  through `python -m surface_morphometrics.xyz2ply`.
- environment.yml / environment-ubuntu.yml install the package via `-e .`;
  Dockerfile copies the package before env creation.
- Archive research/paper/one-off scripts to old_scripts/ and add temporary
  deprecation shims at the repo root for the old `python <script>.py` calls.
- Update README (migration table, install, shell completion) and Install.md.
- config_template.yml: update mesh_refinement defaults to the settings that
  work well (xcorr_iterations hybrid fitting, stronger smoothing, convergence
  threshold).
