+++ old_scripts/distance_to_edge.py
