+++ old_scripts/label_connected_components.py
