+++ old_scripts/mitochondria_statistics.py
