+++ old_scripts/rescale_bad_plys.py
