+++ old_scripts/thick_patch_selector.py
