+++ old_scripts/thickness_stats.py
