+++ old_scripts/thickness_subcompartment.py
