+++ surface_morphometrics/__init__.py
@@ -1,0 +1,7 @@
+"""Surface Morphometrics: quantification of membrane surfaces from cryo-ET.
+
+See https://github.com/GrotjahnLab/surface_morphometrics and the `morphometrics`
+command line interface (defined in `surface_morphometrics.cli`).
+"""
+
+__version__ = "0.2.0"
