+++ surface_morphometrics/_thickness_worker.py
