+++ surface_morphometrics/config_template.yml
@@ -59,20 +59,21 @@ thickness_measurements:
 
 # MESH REFINEMENT (used by refine_mesh.py)
 mesh_refinement:
-  iterations: 5  # Number of refinement iterations
-  damping_factor: 1.0  # Fraction of impulse to apply (0-1). Values 0.5-0.7 work well to prevent oscillation.
+  iterations: 6  # Number of refinement iterations
+  damping_factor: 0.9  # Fraction of impulse to apply (0-1). Lower values prevent oscillation; 0.9 works well.
   monolayer: false  # Set to true for high-defocus data where bilayer is not resolved. Fits single Gaussian instead of dual.
-  use_xcorr: true  # Set to true to use cross-correlation with global average instead of Gaussian fitting. More robust to noise.
-  average_radius: 20  # Radius in nm for local averaging of density profiles. Larger values = more smoothing.
-  average_radius_decay: 0.9  # Factor to multiply radius by each iteration (e.g., 0.8 = 20% reduction). Set to 1.0 for no decay.
+  xcorr_iterations: [1, 2, 3]  # Iterations that use cross-correlation to sharpen the bilayer peaks; the remaining iterations use per-triangle dual Gaussian fitting for precise centering. (Takes priority over use_xcorr; set `use_xcorr: true` instead to use cross-correlation for every iteration.)
+  average_radius: 25  # Radius in nm for local averaging of density profiles. Larger values = more smoothing.
+  average_radius_decay: 1.0  # Factor to multiply radius by each iteration (e.g., 0.8 = 20% reduction). 1.0 = no decay, keeps profiles clean throughout.
   average_radius_min: 12.0  # Minimum averaging radius in nm (prevents over-fitting to noise in later iterations).
-  max_total_offset: 4  # Maximum total displacement from original surface in nm. Prevents divergence in poorly-fit regions.
+  max_total_offset: 8  # Maximum total displacement from original surface in nm. Prevents divergence in poorly-fit regions.
   # Smoothing options to reduce surface roughness
   smooth_offsets: true  # Spatially smooth the offset field before applying to vertices. Reduces local noise.
-  offset_smoothing_radius: 12  # Radius for offset smoothing in nm. If null, uses average_radius.
-  laplacian_iterations: 1  # Number of Laplacian smoothing iterations after displacement (0 = disabled).
-  laplacian_lambda: 0.3  # Laplacian smoothing strength (0-1). Higher = more smoothing but may lose detail.
-  lowpass_sigma: 2  # 3D Gaussian low-pass filter sigma in nm applied to tomogram before sampling. 0 = disabled. Try 1-3 nm for noisy data.
+  offset_smoothing_radius: 25  # Radius for offset smoothing in nm. Match to average_radius to remove noise from the offset field. If null, uses average_radius.
+  laplacian_iterations: 5  # Number of Laplacian smoothing iterations after displacement (0 = disabled).
+  laplacian_lambda: 0.5  # Laplacian smoothing strength (0-1). Higher = more smoothing but may lose detail.
+  lowpass_sigma: 0  # 3D Gaussian low-pass filter sigma in nm applied to tomogram before sampling. 0 = disabled. Try 1-3 nm for noisy data.
+  convergence_threshold: 0.05  # nm - stop iterating (per phase) when mean |offset| drops below this.
